Cleanup unused code, remove tree.go and move useful methods to matcher directory
